@@ -10,13 +10,13 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
-    buildToolsVersion "28.0.3"
+    compileSdkVersion 30
+    buildToolsVersion '30.0.2'
 
     defaultConfig {
         applicationId "com.yandex.admobadapter.banner.sample"
         minSdkVersion 16
-        targetSdkVersion 28
+        targetSdkVersion 30
         versionCode 1
         versionName "1.0"
         multiDexEnabled true
@@ -37,11 +37,11 @@ android {
 
 dependencies {
     implementation 'androidx.multidex:multidex:2.0.1'
-    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'androidx.appcompat:appcompat:1.2.0'
 
-    implementation 'com.yandex.android:mobmetricalib:3.16.2'
-    implementation 'com.yandex.android:mobileads:3.3.0'
-    implementation 'com.yandex.ads.adapter:admob-mobileads:0.6.0'
+    implementation 'com.yandex.android:mobmetricalib:3.18.0'
+    implementation 'com.yandex.android:mobileads:4.0.0'
+    implementation 'com.yandex.ads.adapter:admob-mobileads:0.7.0'
 
-    implementation 'com.google.android.gms:play-services-ads:19.3.0'
+    implementation 'com.google.android.gms:play-services-ads:19.5.0'
 }
