@@ -1,11 +1,11 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 30
     defaultConfig {
         applicationId "com.yandex.admobadapter.nativeads.sample"
         minSdkVersion 16
-        targetSdkVersion 28
+        targetSdkVersion 30
         versionCode 1
         versionName "1.0"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
@@ -26,11 +26,11 @@ android {
 
 dependencies {
     implementation 'androidx.multidex:multidex:2.0.1'
-    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'androidx.appcompat:appcompat:1.2.0'
 
-    implementation 'com.yandex.android:mobmetricalib:3.16.2'
-    implementation 'com.yandex.android:mobileads:3.3.0'
-    implementation 'com.yandex.ads.adapter:admob-mobileads:0.6.0'
+    implementation 'com.yandex.android:mobmetricalib:3.18.0'
+    implementation 'com.yandex.android:mobileads:4.0.0'
+    implementation 'com.yandex.ads.adapter:admob-mobileads:0.7.0'
 
-    implementation 'com.google.android.gms:play-services-ads:19.3.0'
+    implementation 'com.google.android.gms:play-services-ads:19.5.0'
 }
