@@ -8,7 +8,7 @@
  */
 package com.yandex.videoad.sample.adapter;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import com.yandex.mobile.ads.video.models.ad.Creative;
 import com.yandex.mobile.ads.video.models.ad.MediaFile;
