@@ -12,8 +12,8 @@ import android.util.Log;
 
 import androidx.multidex.MultiDexApplication;
 
-import com.yandex.mobile.ads.InitializationListener;
-import com.yandex.mobile.ads.MobileAds;
+import com.yandex.mobile.ads.common.InitializationListener;
+import com.yandex.mobile.ads.common.MobileAds;
 
 public class Application extends MultiDexApplication {
 
