@@ -16,24 +16,24 @@ import android.widget.Toast;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
-import com.yandex.mobile.ads.AdEventListener;
-import com.yandex.mobile.ads.AdRequest;
-import com.yandex.mobile.ads.AdRequestError;
-import com.yandex.mobile.ads.AdSize;
-import com.yandex.mobile.ads.AdView;
+import com.yandex.mobile.ads.banner.AdSize;
+import com.yandex.mobile.ads.banner.BannerAdEventListener;
+import com.yandex.mobile.ads.banner.BannerAdView;
+import com.yandex.mobile.ads.common.AdRequest;
+import com.yandex.mobile.ads.common.AdRequestError;
 
 public class MainActivity extends AppCompatActivity {
 
     private static final String DEMO_BLOCK_ID = "R-M-DEMO-adaptive-sticky";
 
-    private AdView mAdView;
+    private BannerAdView mBannerAdView;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        mAdView = findViewById(R.id.ad_view);
+        mBannerAdView = findViewById(R.id.ad_view);
         initAdView();
 
         final Button loadBannerButton = findViewById(R.id.load_banner_button);
@@ -42,11 +42,11 @@ public class MainActivity extends AppCompatActivity {
 
     private void initAdView() {
         // You can use width of your adview container instead of AdSize.FULL_WIDTH
-        mAdView.setAdSize(AdSize.stickySize(AdSize.FULL_WIDTH));
+        mBannerAdView.setAdSize(AdSize.stickySize(AdSize.FULL_WIDTH));
 
         // Replace demo R-M-DEMO-adaptive-sticky with actual Block ID
-        mAdView.setBlockId(DEMO_BLOCK_ID);
-        mAdView.setAdEventListener(new StickyBannerEventListener());
+        mBannerAdView.setBlockId(DEMO_BLOCK_ID);
+        mBannerAdView.setBannerAdEventListener(new StickyBannerEventListener());
     }
 
 
@@ -57,21 +57,32 @@ public class MainActivity extends AppCompatActivity {
         }
 
         private void refreshBannerAd() {
-            final AdRequest adRequest = AdRequest.builder().build();
-            mAdView.loadAd(adRequest);
+            final AdRequest adRequest = new AdRequest.Builder().build();
+            mBannerAdView.loadAd(adRequest);
         }
     }
 
-    private class StickyBannerEventListener extends AdEventListener.SimpleAdEventListener {
+    private class StickyBannerEventListener implements BannerAdEventListener {
+
         @Override
         public void onAdLoaded() {
             Toast.makeText(MainActivity.this, "Ad loaded", Toast.LENGTH_SHORT).show();
         }
 
         @Override
-        public void onAdFailedToLoad(@NonNull final AdRequestError error) {
+        public void onAdFailedToLoad(@NonNull final AdRequestError adRequestError) {
             Toast.makeText(MainActivity.this, "Failed to load with reason: " +
-                    error.getDescription(), Toast.LENGTH_SHORT).show();
+                    adRequestError.getDescription(), Toast.LENGTH_SHORT).show();
+        }
+
+        @Override
+        public void onLeftApplication() {
+            Toast.makeText(MainActivity.this, "onLeftApplication", Toast.LENGTH_SHORT).show();
+        }
+
+        @Override
+        public void onReturnedToApplication() {
+            Toast.makeText(MainActivity.this, "onReturnedToApplication", Toast.LENGTH_SHORT).show();
         }
     }
 }
