@@ -8,12 +8,13 @@
  */
 package com.yandex.videoad.sample.adapter;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.yandex.ads.videoad.sample.R;
 import com.yandex.mobile.ads.video.models.common.Extension;
